REF: rename gam.py to generalized_additive_model
