+++ statsmodels/gam/api.py
@@ -1,5 +1,5 @@
 # pylint: disable=W0611
 
-from .gam import GLMGam
+from .generalized_additive_model import GLMGam
 from .gam_cross_validation.gam_cross_validation import MultivariateGAMCVPath
 from .smooth_basis import BSplines, CyclicCubicSplines
