+++ statsmodels/gam/generalized_additive_model.py
