+++ statsmodels/gam/tests/test_gam.py
@@ -19,8 +19,8 @@ from statsmodels.tools.linalg import matrix_sqrt
 from statsmodels.gam.smooth_basis import (UnivariatePolynomialSmoother,
     PolynomialSmoother, BSplines, GenericSmoothers, UnivariateCubicSplines,
     CyclicCubicSplines)
-from statsmodels.gam.gam import (GLMGam, LogitGam, make_augmented_matrix,
-                                 penalized_wls)
+from statsmodels.gam.generalized_additive_model import (
+    GLMGam, LogitGam, make_augmented_matrix, penalized_wls)
 from statsmodels.gam.gam_cross_validation.gam_cross_validation import (
     MultivariateGAMCV, MultivariateGAMCVPath, _split_train_test_smoothers)
 from statsmodels.gam.gam_penalties import (UnivariateGamPenalty,
